+++ resource/js/components/PageEditor/MarkdownTableUtil.js
@@ -1,5 +1,3 @@
-import * as codemirror from 'codemirror';
-
 /**
  * Utility for markdown table
  */
